+++ assignment-4/index.ts
@@ -1,11 +1,13 @@
 // Requiring Modules
-const express = require('express');
+import express from 'express';
 const mongoose = require('mongoose');
 var expressLayouts = require('express-ejs-layouts');
+
 const app = express();
+import UserDetails from './models/userDetails';
+
 const passport = require('passport');
 const session = require('express-session');
-const UserDetails = require('./models/userDetails');
 const routes = require('./routes/router');
 require('dotenv').config();
 
@@ -35,13 +37,13 @@ app.use(express.json());
 app.use(passport.initialize());
 app.use(passport.session());
 
-passport.use(UserDetails.createStrategy());
-passport.serializeUser(UserDetails.serializeUser());
-passport.deserializeUser(UserDetails.deserializeUser());
+passport.use((UserDetails as any).createStrategy());
+passport.serializeUser((UserDetails as any).serializeUser());
+passport.deserializeUser((UserDetails as any).deserializeUser());
 
 app.use(routes);
 
 // Set up Express server
 const server = app.listen(3000, () => {
-  console.log(`Listening on port ${server.address().port}`);
+  console.log(`Listening on port 3000`);
 });
