--- assignment-4/models/Notes.js
@@ -1,10 +1,0 @@
-const mongoose = require('mongoose');
-
-const Note = new mongoose.Schema({
-  title: String,
-  createdBy: String,
-  status: String,
-  data: String,
-});
-
-module.exports = mongoose.model('Note', Note);
