+++ assignment-4/models/userDetails.ts
@@ -1,8 +1,12 @@
-const mongoose = require('mongoose');
+import mongoose from 'mongoose';
 const passportLocalMongoose = require('passport-local-mongoose');
 
+export interface IUserDetails {
+  username: string;
+  password: string;
+}
 // Setting up the schema
-const User = new mongoose.Schema({
+const User = new mongoose.Schema<IUserDetails>({
   username: String,
   password: String,
 });
@@ -10,4 +14,4 @@ const User = new mongoose.Schema({
 // Setting up the passport plugin
 User.plugin(passportLocalMongoose);
 
-module.exports = mongoose.model('User', User);
+export default mongoose.model('User', User);
