+++ assignment-4/package.json
@@ -4,8 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
+    "start": "node dist/index.js",
+    "dev": "nodemon dist/index.js",
+    "build": "tsc -p tsconfig.json"
   },
   "keywords": [],
   "author": "",
@@ -23,6 +24,9 @@
     "passport-local-mongoose": "^7.1.2"
   },
   "devDependencies": {
-    "nodemon": "^2.0.18"
+    "@types/express": "^4.17.13",
+    "@types/node": "^18.0.0",
+    "nodemon": "^2.0.18",
+    "typescript": "^4.7.4"
   }
 }
